fix admin user name with refactor gateway
