--- client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/local/gateway/LocalGateway.kt
@@ -1,13 +1,0 @@
-package org.thechance.common.data.local.gateway
-
-
-import org.thechance.common.domain.getway.ILocalGateway
-import java.io.File
-
-class LocalGateway : ILocalGateway {
-
-    override suspend fun saveTaxiReport(file: File) {
-        //todo save file
-    }
-
-}
