+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/local/gateway/UserLocalGateway.kt
@@ -5,9 +5,9 @@ import io.realm.kotlin.ext.query
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import org.thechance.common.data.local.model.ConfigurationCollection
-import org.thechance.common.domain.getway.IIdentityGateway
+import org.thechance.common.domain.getway.IUserLocalGateway
 
-class IdentityGateway(private val realm: Realm) : IIdentityGateway {
+class UserLocalGateway(private val realm: Realm) : IUserLocalGateway {
 
     override suspend fun createUserConfiguration() {
         realm.write { copyToRealm(ConfigurationCollection().apply { id = CONFIGURATION_ID }) }
@@ -22,6 +22,16 @@ class IdentityGateway(private val realm: Realm) : IIdentityGateway {
             query<ConfigurationCollection>("$ID == $CONFIGURATION_ID").first().find()?.accessToken = token
         }
     }
+    override suspend fun saveUserName(username: String) {
+        realm.write {
+            query<ConfigurationCollection>("$ID == $CONFIGURATION_ID").first().find()?.username = username
+        }
+    }
+
+    override suspend fun getUserName(): String {
+        return realm.query<ConfigurationCollection>("$ID == $CONFIGURATION_ID").first().find()?.username ?: ""
+    }
+
 
     override suspend fun saveRefreshToken(token: String) {
         realm.write {
@@ -40,19 +50,6 @@ class IdentityGateway(private val realm: Realm) : IIdentityGateway {
     override suspend fun clearConfiguration() {
         realm.write { delete(query<ConfigurationCollection>()) }
     }
-
-    override suspend fun shouldUserKeptLoggedIn(keepLoggedIn: Boolean) {
-        realm.write {
-            query<ConfigurationCollection>("$ID == $CONFIGURATION_ID").first().find()?.keepLoggedIn = keepLoggedIn
-        }
-    }
-
-    override suspend fun isUserKeptLoggedIn(): Boolean {
-        return realm.query<ConfigurationCollection>(
-            "$ID == $CONFIGURATION_ID"
-        ).first().find()?.keepLoggedIn ?: false
-    }
-
     override suspend fun getThemeMode(): Flow<Boolean> {
         return realm.query<ConfigurationCollection>(
             "$ID == $CONFIGURATION_ID"
