+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/local/model/ConfigurationCollection.kt
@@ -4,7 +4,7 @@ import io.realm.kotlin.types.RealmObject
 
 class ConfigurationCollection : RealmObject {
     var id: Int = 0
-    var keepLoggedIn: Boolean = false
+    var username: String = ""
     var accessToken: String = ""
     var refreshToken: String = ""
     var isDarkMode: Boolean = false
