+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/remote/fakegateway/TaxisFakeGateway.kt
@@ -1,5 +1,10 @@
 package org.thechance.common.data.remote.fakegateway
 
+import org.apache.pdfbox.pdmodel.PDDocument
+import org.apache.pdfbox.pdmodel.PDPage
+import org.apache.pdfbox.pdmodel.PDPageContentStream
+import org.apache.pdfbox.pdmodel.common.PDRectangle
+import org.apache.pdfbox.pdmodel.font.PDType1Font
 import org.thechance.common.data.remote.mapper.toDto
 import org.thechance.common.data.remote.mapper.toEntity
 import org.thechance.common.data.remote.model.DataWrapperDto
@@ -10,10 +15,16 @@ import org.thechance.common.domain.entity.NewTaxiInfo
 import org.thechance.common.domain.entity.Taxi
 import org.thechance.common.domain.entity.TaxiFiltration
 import org.thechance.common.domain.getway.ITaxisGateway
+import java.io.File
+import java.text.SimpleDateFormat
 import java.util.*
 import kotlin.math.ceil
 
 class TaxisFakeGateway : ITaxisGateway {
+
+    override suspend fun getPdfTaxiReport() {
+        createTaxiPDFReport()
+    }
 
     override suspend fun getTaxis(
         username: String?,
@@ -346,5 +357,185 @@ class TaxisFakeGateway : ITaxisGateway {
             driverId = "DRIVER_ID"
         ),
     )
+
+    private fun createTaxiPDFReport(): File {
+        val title = "Taxi Details Report"
+        val columnNames = listOf(
+                "Taxi ID",
+                "Username",
+                "Plate Number",
+                "Model",
+                "Color",
+                "Seats",
+                "Status",
+                "Trips"
+        )
+        val columnWidth = listOf(50f, 80f, 80f, 80f, 80f, 80f, 80f, 50f)
+
+        val file = createPDFReport(title, taxis, columnNames, columnWidth) { taxi ->
+            listOf(
+                    taxi.id.toString(), taxi.username, taxi.plateNumber, taxi.type, taxi.color,
+                    taxi.seats, taxi.isAvailable.toString(), taxi.trips.toString()
+            ).map { it ?: "" }
+        }
+        return file
+    }
+
+    private fun <T> createPDFReport(
+        title: String,
+        dataList: List<T>,
+        columnNames: List<String>,
+        colWidths: List<Float>,
+        dataExtractor: (T) -> List<Any>,
+    ): File {
+        try {
+            //region Create PDF document
+            val document = PDDocument()
+            val page = PDPage(PDRectangle.A4)
+            document.addPage(page)
+            val contentStream = PDPageContentStream(document, page)
+
+            val pageWidth = page.trimBox.width
+            val pageHeight = page.trimBox.height
+
+            val startX = 10f
+            val cellHeight = 30f
+            //endregion
+            val titleY = titleContent(contentStream, title, pageWidth, pageHeight)
+            val dateTimeY = dateContent(contentStream, pageWidth, titleY)
+            val (currentX, currentY) = headerContent(
+                    contentStream,
+                    startX,
+                    dateTimeY,
+                    cellHeight,
+                    columnNames,
+                    colWidths
+            )
+            tableContent(
+                    currentY,
+                    contentStream,
+                    dataList,
+                    cellHeight,
+                    currentX,
+                    startX,
+                    colWidths,
+                    dataExtractor
+            )
+            contentStream.close()
+            val pdfFilePath = "${System.getProperty("user.home")}/Downloads/$title.pdf"
+            val pdfFile = File(pdfFilePath)
+            document.save(pdfFile)
+            document.close()
+            return pdfFile
+        } catch (e: Exception) {
+            e.printStackTrace()
+            return File("")
+        }
+    }
+
+    private fun <T> tableContent(
+        currentY: Float,
+        contentStream: PDPageContentStream,
+        dataList: List<T>,
+        cellHeight: Float,
+        currentX: Float,
+        startX: Float,
+        colWidths: List<Float>,
+        dataExtractor: (T) -> List<Any>,
+    ) {
+        // Draw table content
+        // Add space between header and table content
+        var currentY = currentY
+        var currentX = currentX
+        val contentTopMargin = 30f // Add the desired space
+        currentY -= contentTopMargin
+        contentStream.setFont(PDType1Font.HELVETICA, 12f)
+        for ((rowIndex, item) in dataList.withIndex()) {
+            currentY -= cellHeight // Move down for the next row
+            currentX = startX
+
+            val rowData = dataExtractor(item)
+
+            for ((index, cellData) in rowData.withIndex()) {
+                contentStream.beginText()
+                val textWidth =
+                    PDType1Font.HELVETICA.getStringWidth(cellData.toString()) / 1000 * 12f
+                val textX = currentX + (colWidths[index] - textWidth) / 2
+                val textY = currentY + 20f
+                contentStream.newLineAtOffset(textX, textY)
+                contentStream.showText(cellData.toString())
+                contentStream.endText()
+
+                currentX += colWidths[index]
+            }
+        }
+    }
+
+    private fun titleContent(
+        contentStream: PDPageContentStream,
+        title: String,
+        pageWidth: Float,
+        pageHeight: Float,
+    ): Float {
+        contentStream.setFont(PDType1Font.HELVETICA_BOLD, 16f)
+        contentStream.beginText()
+        val titleWidth = PDType1Font.HELVETICA_BOLD.getStringWidth(title) / 1000 * 16f
+        val titleX = pageWidth / 2 - titleWidth / 2
+        val titleY = pageHeight - 30f
+        contentStream.newLineAtOffset(titleX, titleY)
+        contentStream.showText(title)
+        contentStream.endText()
+        return titleY
+    }
+
+    private fun dateContent(
+        contentStream: PDPageContentStream,
+        pageWidth: Float,
+        titleY: Float,
+    ): Float {
+        val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
+        val dateTime = dateFormat.format(Date())
+        contentStream.setFont(PDType1Font.HELVETICA, 11f)
+
+        contentStream.beginText()
+        val dateTimeWidth = PDType1Font.HELVETICA.getStringWidth(dateTime) / 1000 * 12f
+        val pageMidX = pageWidth / 2 // Place the text in the center of the page
+        val dateTimeX = pageMidX - dateTimeWidth / 2 // Place the text in the center horizontally
+        val dateTimeTopMargin = 10f // Add space between date/time and title
+        val dateTimeY = titleY - 20f - dateTimeTopMargin // Adjust the value to add more space
+        contentStream.newLineAtOffset(dateTimeX, dateTimeY)
+        contentStream.showText(dateTime)
+        contentStream.endText()
+        return dateTimeY
+    }
+
+    private fun headerContent(
+        contentStream: PDPageContentStream,
+        startX: Float,
+        dateTimeY: Float,
+        cellHeight: Float,
+        columnNames: List<String>,
+        colWidths: List<Float>,
+    ): Pair<Float, Float> {
+        // region Draw header row
+        contentStream.setFont(PDType1Font.HELVETICA_BOLD, 11f)
+        var currentX = startX
+        val headerTopMargin = 20f // Add space between title/date
+        val headerY = dateTimeY - headerTopMargin - cellHeight
+        var currentY = headerY
+        val columnWidth = columnNames.size
+        for ((index, columnName) in columnNames.withIndex()) {
+            contentStream.beginText()
+            val textWidth = PDType1Font.HELVETICA_BOLD.getStringWidth(columnName) / 1000 * 12f
+            val textX = currentX + (colWidths[index] - textWidth) / 2
+            val textY = currentY - cellHeight + 20f
+            contentStream.newLineAtOffset(textX, textY)
+            contentStream.showText(columnName)
+            contentStream.endText()
+            currentX += colWidths[index]
+        }
+        return Pair(currentX, currentY)
+    }
+
 
 }
