+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/remote/gateway/UsersGateway.kt
@@ -20,8 +20,6 @@ import org.thechance.common.domain.getway.IUsersGateway
 
 class UsersGateway(private val client: HttpClient) : BaseGateway(), IUsersGateway {
 
-    override suspend fun getUserData(): String = "aaaa"
-
     override suspend fun getUsers(
         query: String?,
         byPermissions: List<Permission>,
