--- client_dashboard/common/src/commonMain/kotlin/org/thechance/common/domain/getway/ILocalGateway.kt
@@ -1,10 +1,0 @@
-package org.thechance.common.domain.getway
-
-import java.io.File
-
-
-interface ILocalGateway {
-
-     suspend fun saveTaxiReport(file: File)
-
-}
