+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/domain/getway/IRemoteGateway.kt
@@ -7,13 +7,7 @@ import org.thechance.common.domain.util.RevenueShareDate
 
 interface IRemoteGateway {
 
-    suspend fun getUserData(): String
-
-    suspend fun getPdfTaxiReport()
-
     suspend fun getCurrentLocation(): String
-
-    suspend fun loginUser(username: String, password: String): Pair<String, String>
 
     suspend fun getRevenueShare(revenueShareDate: RevenueShareDate): TotalRevenueShare
     suspend fun getDashboardRevenueShare(): RevenueShare
