+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/domain/getway/ITaxisGateway.kt
@@ -7,6 +7,7 @@ import org.thechance.common.domain.entity.TaxiFiltration
 
 interface ITaxisGateway {
 
+    suspend fun getPdfTaxiReport()
     suspend fun getTaxis(
         username: String?,
         taxiFiltration: TaxiFiltration,
