+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/domain/getway/IUsersGateway.kt
@@ -6,8 +6,6 @@ import org.thechance.common.domain.entity.User
 
 interface IUsersGateway {
 
-    suspend fun getUserData(): String
-
     suspend fun getUsers(
         query: String?,
         byPermissions: List<Permission>,
