+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/domain/usecase/UsersManagementUseCase.kt
@@ -3,6 +3,7 @@ package org.thechance.common.domain.usecase
 import org.thechance.common.domain.entity.DataWrapper
 import org.thechance.common.domain.entity.Permission
 import org.thechance.common.domain.entity.User
+import org.thechance.common.domain.getway.IUserLocalGateway
 import org.thechance.common.domain.getway.IUsersGateway
 
 interface IUsersManagementUseCase {
@@ -25,7 +26,8 @@ interface IUsersManagementUseCase {
 }
 
 class UsersManagementUseCase(
-    private val userGateway: IUsersGateway
+    private val userGateway: IUsersGateway,
+    private val userLocalGateway: IUserLocalGateway
 ) : IUsersManagementUseCase {
 
     override suspend fun getUsers(
@@ -54,7 +56,7 @@ class UsersManagementUseCase(
     }
 
     override suspend fun getUserInfo(): String {
-        return userGateway.getUserData()
+        return userLocalGateway.getUserName()
     }
 
 }
