+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/presentation/login/LoginInteractionListener.kt
@@ -10,6 +10,4 @@ interface LoginInteractionListener : BaseInteractionListener {
 
     fun onLoginClicked()
 
-    fun onKeepLoggedInClicked()
-
 }
