+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/presentation/login/LoginScreenModel.kt
@@ -27,16 +27,11 @@ class LoginScreenModel(
                 login.loginUser(
                     username = currentState.username,
                     password = currentState.password,
-                    keepLoggedIn = currentState.keepLoggedIn
                 )
             },
             onSuccess = { onLoginSuccess() },
             onError = ::onLoginError
         )
-    }
-
-    override fun onKeepLoggedInClicked() {
-        updateState { it.copy(keepLoggedIn = !it.keepLoggedIn) }
     }
 
     private fun onLoginSuccess() {
