+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/presentation/login/LoginUIState.kt
@@ -7,10 +7,9 @@ import org.thechance.common.presentation.util.ErrorState
 data class LoginUIState(
     val username: String = "",
     val password: String = "",
-    val isLoading: Boolean = false,
+    val isLoading: Boolean = true,
     val error: ErrorState? = null,
     val isUserError: ErrorWrapper = ErrorWrapper(),
     val isPasswordError: ErrorWrapper = ErrorWrapper(),
-    val keepLoggedIn: Boolean = false,
     val isAbleToLogin: Boolean = false
 )
