+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/presentation/main/MainContainer.kt
@@ -25,9 +25,7 @@ import org.thechance.common.presentation.resources.Resources
 
 object MainContainer :
     BaseScreen<MainScreenModel, MainUiEffect, MainUiState, MainInteractionListener>() {
-
-    private fun readResolve(): Any = MainContainer
-
+    
     @Composable
     override fun Content() {
         Init(getScreenModel())
