+++ client_dashboard/desktop/src/jvmMain/kotlin/di/GatewayModule.kt
@@ -3,10 +3,8 @@ package di
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
-import org.thechance.common.data.local.gateway.IdentityGateway
-import org.thechance.common.data.local.gateway.LocalGateway
-import org.thechance.common.data.remote.fakegateway.RestaurantFakeGateway
-import org.thechance.common.data.remote.gateway.FakeRemoteGateway
+import org.thechance.common.data.local.gateway.UserLocalGateway
+import org.thechance.common.data.remote.fakegateway.FakeRemoteGateway
 import org.thechance.common.data.remote.gateway.RestaurantGateway
 import org.thechance.common.data.remote.gateway.TaxisGateway
 import org.thechance.common.data.remote.gateway.UsersGateway
@@ -14,14 +12,13 @@ import org.thechance.common.domain.getway.*
 
 val GatewayModule = module {
     // region Real Gateways
-    singleOf(::IdentityGateway) { bind<IIdentityGateway>() }
+    singleOf(::UserLocalGateway) { bind<IUserLocalGateway>() }
     singleOf(::UsersGateway) { bind<IUsersGateway>() }
     singleOf(::TaxisGateway) { bind<ITaxisGateway>() }
     singleOf(::RestaurantGateway){ bind<IRestaurantGateway>() }
     // endregion
 
     // region Fake Gateways
-    singleOf(::LocalGateway) { bind<ILocalGateway>() }
     singleOf(::FakeRemoteGateway) { bind<IRemoteGateway>() }
 //    singleOf(::UsersFakeGateway){ bind<IUsersGateway>()}
 //    singleOf(::TaxisFakeGateway) { bind<ITaxisGateway>() }
